compiles and starts

+++ app/org/intracer/wmua/Comment.scala
@@ -5,9 +5,9 @@ import scalikejdbc._
 
 case class Comment(
   id: Long,
-  userId: Int,
+  userId: Long,
   username: String,
-  round: Int,
+  round: Long,
   room: Long,
   createdAt: String,
   body: String)
@@ -15,7 +15,6 @@ case class Comment(
 
 
 }
-
 
 object CommentJdbc extends SQLSyntaxSupport[Comment] {
 
@@ -35,7 +34,7 @@ object CommentJdbc extends SQLSyntaxSupport[Comment] {
     body = rs.string(c.body)
   )
 
-  def create(userId: Int, username: String, round: Int, room: Long, body: String, createdAt: String = DateTime.now.toString)(implicit session: DBSession = autoSession): Comment = {
+  def create(userId: Long, username: String, round: Long, room: Long, body: String, createdAt: String = DateTime.now.toString)(implicit session: DBSession = autoSession): Comment = {
     val id = withSQL {
       insert.into(CommentJdbc).namedValues(
         column.userId -> userId,
@@ -55,7 +54,7 @@ object CommentJdbc extends SQLSyntaxSupport[Comment] {
       .orderBy(c.id)
   }.map(CommentJdbc(c)).list().apply()
 
-  def findByRound(round: Int)(implicit session: DBSession = autoSession): List[Comment] = withSQL {
+  def findByRound(round: Long)(implicit session: DBSession = autoSession): List[Comment] = withSQL {
     select.from(CommentJdbc as c)
       .where //.append(isNotDeleted)
       // .and
@@ -63,7 +62,7 @@ object CommentJdbc extends SQLSyntaxSupport[Comment] {
       .orderBy(c.id)
   }.map(CommentJdbc(c)).list().apply()
 
-  def findByRoundAndSubject(round: Int, subject: Long)(implicit session: DBSession = autoSession): List[Comment] = withSQL {
+  def findByRoundAndSubject(round: Long, subject: Long)(implicit session: DBSession = autoSession): List[Comment] = withSQL {
     select.from(CommentJdbc as c)
       .where //.append(isNotDeleted)
       // .and
